WFLY-10519 remove EE8 checks for JPA, default to JPA 2.2 + Hibernate ORM 5.3.1.Final + Hibernate Search 5.10.1.Final
