+++ feature-pack/src/main/resources/content/docs/licenses/full-feature-pack-licenses-addendum.html
@@ -36,7 +36,6 @@ org.hibernate:hibernate-search-backend-jms:5.10.1.Final
 org.hibernate:hibernate-search-engine:5.10.1.Final
 org.hibernate:hibernate-search-orm:5.10.1.Final
 org.hibernate:hibernate-search-serialization-avro:5.10.1.Final
-org.hibernate.javax.persistence:hibernate-jpa-2.1-api:1.0.2.Final
 -->
 <html>
 <head>
@@ -160,17 +159,6 @@ org.hibernate.javax.persistence:hibernate-jpa-2.1-api:1.0.2.Final
         <td>org.hibernate</td><td>hibernate-search-serialization-avro</td><td>5.10.1.Final</td><td><a href="http://www.opensource.org/licenses/LGPL-2.1">GNU Lesser General Public License v2.1 or later</a>
         <br>
     </td><td><a href="gnu lesser general public license v2.1 or later.txt">gnu lesser general public license v2.1 or later.txt</a>
-        <br>
-    </td>
-    </tr>
-    <tr>
-        <td>org.hibernate.javax.persistence</td><td>hibernate-jpa-2.1-api</td><td>1.0.2.Final</td><td><a href="http://repository.jboss.org/licenses/edl-1.0.txt">Eclipse Distribution License, Version 1.0</a>
-        <br>
-        <a href="http://repository.jboss.org/licenses/epl-1.0.txt">Eclipse Public License 1.0</a>
-        <br>
-    </td><td><a href="eclipse distribution license, version 1.0.txt">eclipse distribution license, version 1.0.txt</a>
-        <br>
-        <a href="eclipse public license 1.0.txt">eclipse public license 1.0.txt</a>
         <br>
     </td>
     </tr>
