+++ feature-pack/src/main/resources/modules/system/layers/base/javax/persistence/api/main/module.xml
@@ -29,15 +29,6 @@
     </dependencies>
 
     <resources>
-        <artifact name="${javax.persistence:javax.persistence-api}">
-            <conditions>
-                <property-equal name="ee8.preview.mode" value="true" />
-            </conditions>
-        </artifact>
-        <artifact name="org.hibernate.javax.persistence:hibernate-jpa-2.1-api:1.0.2.Final"> <!-- ${version.org.hibernate.javax.persistence.hibernate-jpa-2.1-api} -->
-            <conditions>
-                <property-not-equal name="ee8.preview.mode" value="true" />
-            </conditions>
-        </artifact>
+        <artifact name="${javax.persistence:javax.persistence-api}"/>
     </resources>
 </module>
