--- feature-pack/src/main/resources/modules/system/layers/base/org/apache/lucene/5.5/module.xml
@@ -1,43 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<!--
-  ~ JBoss, Home of Professional Open Source.
-  ~ Copyright 2018, Red Hat, Inc., and individual contributors
-  ~ as indicated by the @author tags. See the copyright.txt file in the
-  ~ distribution for a full listing of individual contributors.
-  ~
-  ~ This is free software; you can redistribute it and/or modify it
-  ~ under the terms of the GNU Lesser General Public License as
-  ~ published by the Free Software Foundation; either version 2.1 of
-  ~ the License, or (at your option) any later version.
-  ~
-  ~ This software is distributed in the hope that it will be useful,
-  ~ but WITHOUT ANY WARRANTY; without even the implied warranty of
-  ~ MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU
-  ~ Lesser General Public License for more details.
-  ~
-  ~ You should have received a copy of the GNU Lesser General Public
-  ~ License along with this software; if not, write to the Free
-  ~ Software Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA
-  ~ 02110-1301 USA, or see the FSF site: http://www.fsf.org.
-  -->
-
-<module xmlns="urn:jboss:module:1.7" name="org.apache.lucene:5.5">
-    <properties>
-        <property name="jboss.api" value="private"/>
-    </properties>
-
-    <resources>
-        <artifact name="org.apache.lucene:lucene-analyzers-common:5.5.5"/> <!-- ${version.org.apache.lucene55} -->
-        <artifact name="org.apache.lucene:lucene-core:5.5.5"/> <!-- ${version.org.apache.lucene55} -->
-        <artifact name="org.apache.lucene:lucene-facet:5.5.5"/> <!-- ${version.org.apache.lucene55} -->
-        <artifact name="org.apache.lucene:lucene-queries:5.5.5"/> <!-- ${version.org.apache.lucene55} -->
-        <artifact name="org.apache.lucene:lucene-queryparser:5.5.5"/> <!-- ${version.org.apache.lucene55} -->
-        <artifact name="org.apache.lucene:lucene-grouping:5.5.5"/> <!-- ${version.org.apache.lucene55} -->
-    </resources>
-
-    <dependencies>
-        <module name="javax.api"/>
-        <module name="org.apache.lucene.internal:5.5" optional="true" services="import"/>
-    </dependencies>
-</module>
