+++ feature-pack/src/main/resources/modules/system/layers/base/org/hibernate/5.3/module.xml
@@ -1,7 +1,8 @@
 <?xml version="1.0" encoding="UTF-8"?>
+
 <!--
   ~ JBoss, Home of Professional Open Source.
-  ~ Copyright 2018, Red Hat, Inc., and individual contributors
+  ~ Copyright 2011, Red Hat, Inc., and individual contributors
   ~ as indicated by the @author tags. See the copyright.txt file in the
   ~ distribution for a full listing of individual contributors.
   ~
@@ -21,18 +22,6 @@
   ~ 02110-1301 USA, or see the FSF site: http://www.fsf.org.
   -->
 
-<module xmlns="urn:jboss:module:1.7" name="org.hibernate.search.backend-jms:5.10">
-    <properties>
-        <property name="jboss.api" value="private"/>
-    </properties>
-    <resources>
-        <artifact name="org.hibernate:hibernate-search-backend-jms:5.10.1.Final"/> <!-- ${version.org.hibernate.search5_10} -->
-    </resources>
-    <dependencies>
-        <module name="org.hibernate.search.engine:5.10"/>
-        <module name="org.jboss.logging"/>
-        <module name="javax.jms.api"/>
-        <module name="org.jboss.as.naming"/>
-        <module name="javax.api"/>
-    </dependencies>
-</module>
+<!-- Represents the Hibernate 5.3.x module-->
+<module-alias xmlns="urn:jboss:module:1.5" name="org.hibernate" slot="5.3"  target-name="org.hibernate"/>
+
