+++ feature-pack/src/main/resources/modules/system/layers/base/org/hibernate/jipijapa-hibernate5-3/main/module.xml
@@ -2,7 +2,7 @@
 
 <!--
   ~ JBoss, Home of Professional Open Source.
-  ~ Copyright 2018, Red Hat, Inc., and individual contributors
+  ~ Copyright 2015, Red Hat, Inc., and individual contributors
   ~ as indicated by the @author tags. See the copyright.txt file in the
   ~ distribution for a full listing of individual contributors.
   ~
@@ -22,18 +22,31 @@
   ~ 02110-1301 USA, or see the FSF site: http://www.fsf.org.
   -->
 
-<module xmlns="urn:jboss:module:1.7" name="org.apache.lucene.internal:5.5">
+<module xmlns="urn:jboss:module:1.7" name="org.hibernate.jipijapa-hibernate5-3">
     <properties>
         <property name="jboss.api" value="private"/>
     </properties>
 
     <resources>
-        <artifact name="org.apache.lucene:lucene-misc:5.5.5"/> <!-- ${version.org.apache.lucene55} -->
-        <artifact name="org.apache.lucene:lucene-backward-codecs:5.5.5"/> <!-- ${version.org.apache.lucene55} -->
+        <artifact name="${org.wildfly:jipijapa-hibernate5-3}"/>
+        <artifact name="${org.wildfly:jipijapa-hibernate5-3-legacy}"/>
     </resources>
 
     <dependencies>
+        <module name="org.hibernate"/>
         <module name="javax.api"/>
-        <module name="org.apache.lucene:5.5"/>
+        <module name="javax.annotation.api"/>
+        <module name="javax.enterprise.api"/>
+        <module name="javax.persistence.api"/>
+        <module name="javax.transaction.api"/>
+        <module name="javax.validation.api"/>
+        <module name="javax.xml.bind.api"/>
+        <module name="org.jboss.as.jpa.spi"/>
+        <module name="org.jboss.jandex"/>
+        <module name="org.jboss.logging"/>
+        <module name="org.jboss.vfs"/>
+        <module name="org.hibernate.commons-annotations"/>
+        <module name="org.infinispan" services="import"/>
+        <module name="org.infinispan.hibernate-cache" services="import"/>
     </dependencies>
 </module>
