+++ feature-pack/src/main/resources/modules/system/layers/base/org/hibernate/jipijapa-hibernate5/main/module.xml
@@ -28,27 +28,8 @@
     </properties>
 
     <resources>
-        <artifact name="${org.wildfly:jipijapa-hibernate5}">
-            <conditions>
-              <property-not-equal name="ee8.preview.mode" value="true"/>
-            </conditions>
-        </artifact>
-        <artifact name="${org.wildfly:jipijapa-hibernate5-3}">
-            <conditions>
-              <property-equal name="ee8.preview.mode" value="true"/>
-            </conditions>
-        </artifact>
-
-        <artifact name="${org.wildfly:jipijapa-hibernate5-legacy}">
-            <conditions>
-              <property-not-equal name="ee8.preview.mode" value="true"/>
-            </conditions>
-        </artifact>
-        <artifact name="${org.wildfly:jipijapa-hibernate5-3-legacy}">
-            <conditions>
-              <property-equal name="ee8.preview.mode" value="true"/>
-            </conditions>
-        </artifact>
+        <artifact name="${org.wildfly:jipijapa-hibernate5}"/>
+        <artifact name="${org.wildfly:jipijapa-hibernate5-legacy}"/>
     </resources>
 
     <dependencies>
