+++ feature-pack/src/main/resources/modules/system/layers/base/org/hibernate/main/module.xml
@@ -25,39 +25,8 @@
 <!-- Represents Hibernate 5.1.x, will later become 5.3.x for WF 14  -->
 <module xmlns="urn:jboss:module:1.7" name="org.hibernate">
     <resources>
-        <artifact name="org.hibernate:hibernate-core:5.3.1.Final"> <!-- ${version.org.hibernate53} -->
-            <conditions>
-              <property-equal name="ee8.preview.mode" value="true"/>
-            </conditions>
-        </artifact>
-        <artifact name="${org.hibernate:hibernate-core}"> 
-            <conditions>
-              <property-not-equal name="ee8.preview.mode" value="true"/>
-            </conditions>
-        </artifact>
-        <artifact name="org.hibernate:hibernate-envers:5.3.1.Final">  <!-- ${version.org.hibernate53} -->
-            <conditions>
-              <property-equal name="ee8.preview.mode" value="true"/>
-            </conditions>
-        </artifact>
-        <artifact name="${org.hibernate:hibernate-envers}"> 
-            <conditions>
-              <property-not-equal name="ee8.preview.mode" value="true"/>
-            </conditions>
-        </artifact>
-
-        <artifact name="${org.hibernate:hibernate-entitymanager}"> 
-            <conditions>
-              <property-not-equal name="ee8.preview.mode" value="true"/>
-            </conditions>
-        </artifact>    
-
-        <artifact name="${org.hibernate:hibernate-java8}"> 
-            <conditions>
-              <property-not-equal name="ee8.preview.mode" value="true"/>
-            </conditions>
-        </artifact>
-
+        <artifact name="${org.hibernate:hibernate-core}"/>
+        <artifact name="${org.hibernate:hibernate-envers}"/> 
     </resources>
 
     <dependencies>
@@ -77,7 +46,7 @@
         <module name="org.jboss.logging"/>
         <module name="org.jboss.vfs"/>
         <module name="org.hibernate.commons-annotations"/>
-        <module name="org.hibernate.jipijapa-hibernate5" services="import"/>
+        <module name="org.hibernate.jipijapa-hibernate5-3" services="import"/>
         <module name="org.infinispan.hibernate-cache" services="import" optional="true"/>
         <module name="net.bytebuddy"/>
     </dependencies>
