--- feature-pack/src/main/resources/modules/system/layers/base/org/hibernate/search/engine/5.10/module.xml
@@ -1,65 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<!--
-  ~ JBoss, Home of Professional Open Source.
-  ~ Copyright 2018, Red Hat, Inc., and individual contributors
-  ~ as indicated by the @author tags. See the copyright.txt file in the
-  ~ distribution for a full listing of individual contributors.
-  ~
-  ~ This is free software; you can redistribute it and/or modify it
-  ~ under the terms of the GNU Lesser General Public License as
-  ~ published by the Free Software Foundation; either version 2.1 of
-  ~ the License, or (at your option) any later version.
-  ~
-  ~ This software is distributed in the hope that it will be useful,
-  ~ but WITHOUT ANY WARRANTY; without even the implied warranty of
-  ~ MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU
-  ~ Lesser General Public License for more details.
-  ~
-  ~ You should have received a copy of the GNU Lesser General Public
-  ~ License along with this software; if not, write to the Free
-  ~ Software Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA
-  ~ 02110-1301 USA, or see the FSF site: http://www.fsf.org.
-  -->
-
-<!-- Hibernate Search Engine: the fulltext indexing and query capabilities
-     without the ORM integration bits. This is reused by several other project
-     outside of the Hibernate group. -->
-<module xmlns="urn:jboss:module:1.7" name="org.hibernate.search.engine:5.10">
-    <properties>
-        <property name="jboss.api" value="private"/>
-    </properties>
-    <resources>
-        <artifact name="org.hibernate:hibernate-search-engine:5.10.1.Final"/>  <!-- ${version.org.hibernate.search5_10} -->
-    </resources>
-
-    <dependencies>
-        <module name="javax.transaction.api"/>
-        <module name="org.hibernate.commons-annotations"/>
-        <module name="org.apache.lucene:5.5" export="true"/>
-        <module name="org.apache.lucene.internal:5.5"/>
-        <module name="org.jboss.logging"/>
-
-        <!-- Elasticsearch support -->
-        <module name="org.hibernate.search.elasticsearch:5.10" export="true" services="import" optional="true" />
-
-        <!-- optional JMS backend support -->
-        <module name="org.hibernate.search.backend-jms:5.10" export="true" services="import" optional="true"/>
-
-        <!-- optional JGroups backend support -->
-        <module name="org.hibernate.search.backend-jgroups:5.10" export="true" services="import" optional="true"/>
-
-        <!-- optional Avro serialization -->
-        <module name="org.hibernate.search.serialization-avro:5.10" services="import" optional="true"/>
-
-        <!-- Infinispan extensions - provided as separate download by the Infinispan project -->
-        <module name="org.infinispan.hibernate-search.directory-provider:for-hibernatesearch-5.10" services="import" optional="true"/>
-
-        <!-- optional Apache Tika integration -->
-        <module name="org.apache.tika" optional="true" />
-
-        <!-- For naming -->
-        <module name="javax.api" />
-        <module name="org.jboss.as.naming" />
-    </dependencies>
-</module>
