+++ feature-pack/src/main/resources/modules/system/layers/base/org/hibernate/search/orm/main/module.xml
@@ -24,7 +24,7 @@
 <!-- Hibernate Search ORM: integrates org.hibernate.search.engine with
      Hibernate ORM to provide Hibernate Search functionality to
      JPA Applications -->
-<module xmlns="urn:jboss:module:1.5" name="org.hibernate.search.orm">
+<module xmlns="urn:jboss:module:1.7" name="org.hibernate.search.orm">
     <resources>
         <artifact name="${org.hibernate:hibernate-search-orm}"/>
     </resources>
@@ -36,5 +36,6 @@
         <module name="org.hibernate.search.engine" export="true" services="import" />
         <module name="org.jboss.logging" />
         <module name="javax.persistence.api" />
+        <module name="javax.enterprise.api"/>
     </dependencies>
 </module>
