+++ feature-pack/src/main/resources/modules/system/layers/base/org/infinispan/hibernate-cache/main/module.xml
@@ -32,23 +32,14 @@
     <resources>
         <artifact name="${org.infinispan:infinispan-hibernate-cache-spi}"/>
         <artifact name="${org.infinispan:infinispan-hibernate-cache-commons}"/>
-        <artifact name="${org.infinispan:infinispan-hibernate-cache-v53}">
-            <conditions>
-              <property-equal name="ee8.preview.mode" value="true"/>
-            </conditions>
-        </artifact>
-        <artifact name="${org.infinispan:infinispan-hibernate-cache-v51}">
-            <conditions>
-              <property-not-equal name="ee8.preview.mode" value="true"/>
-            </conditions>
-        </artifact>
+        <artifact name="${org.infinispan:infinispan-hibernate-cache-v53}"/>
     </resources>
 
     <dependencies>
         <module name="org.hibernate"/>
         <module name="javax.api"/>
         <module name="javax.transaction.api"/>
-        <module name="org.hibernate.jipijapa-hibernate5" services="import"/>
+        <module name="org.hibernate.jipijapa-hibernate5-3" services="import"/>
         <module name="org.infinispan" services="import"/>
         <module name="org.infinispan.commons"/>
         <module name="org.jboss.as.jpa.spi"/>
