+++ jpa/subsystem/src/main/java/org/jboss/as/jpa/puparser/PersistenceUnitXmlParser.java
@@ -162,8 +162,10 @@ public class PersistenceUnitXmlParser extends MetaDataElementParser {
             pu.setPersistenceXMLSchemaVersion("1.0");
         } else if (version.equals(Version.JPA_2_0)) {
             pu.setPersistenceXMLSchemaVersion("2.0");
-        } else {
+        } else if (version.equals(Version.JPA_2_1)) {
             pu.setPersistenceXMLSchemaVersion("2.1");
+        } else {
+            pu.setPersistenceXMLSchemaVersion("2.2");
         }
 
         final int count = reader.getAttributeCount();
