+++ pom.xml
@@ -250,8 +250,7 @@
         <version.org.apache.httpcomponents.httpasyncclient>4.1.3</version.org.apache.httpcomponents.httpasyncclient>
         <version.org.apache.james.apache-mime4j>0.6</version.org.apache.james.apache-mime4j>
         <version.org.apache.jstl>1.2.6-RC1</version.org.apache.jstl>
-        <version.org.apache.lucene>5.3.1</version.org.apache.lucene>
-        <version.org.apache.lucene55>5.5.5</version.org.apache.lucene55>
+        <version.org.apache.lucene>5.5.5</version.org.apache.lucene>
         <version.org.apache.myfaces.core>2.2.11</version.org.apache.myfaces.core>
         <version.org.apache.neethi>3.1.1</version.org.apache.neethi>
         <version.org.apache.openjpa>2.4.2</version.org.apache.openjpa>
@@ -274,16 +273,10 @@
         <version.org.glassfish.javax.enterprise.concurrent>1.0</version.org.glassfish.javax.enterprise.concurrent>
         <version.org.glassfish.javax.json>1.1.2</version.org.glassfish.javax.json>
         <version.org.glassfish.soteria>1.0</version.org.glassfish.soteria>
-        <version.org.hibernate53>5.3.1.Final</version.org.hibernate53>
-        <!-- Hardcoded in module.xml for the dual EE7 and EE8 support -->
-        <version.org.hibernate>5.1.14.Final</version.org.hibernate>
+        <version.org.hibernate>5.3.2.Final</version.org.hibernate>
         <version.org.hibernate.commons.annotations>5.0.4.Final</version.org.hibernate.commons.annotations>
-        <version.org.hibernate.search>5.5.8.Final</version.org.hibernate.search>
-        <!-- Hardcoded in module.xml for the dual EE7 and EE8 support -->
-        <version.org.hibernate.search5_10>5.10.1.Final</version.org.hibernate.search5_10>
+        <version.org.hibernate.search>5.10.1.Final</version.org.hibernate.search>        <version.org.hibernate.validator>6.0.10.Final</version.org.hibernate.validator>
         <version.org.hibernate.validator>6.0.10.Final</version.org.hibernate.validator>
-        <!-- Hardcoded in module.xml for the dual EE7 and EE8 support -->
-        <version.org.hibernate.javax.persistence.hibernate-jpa-2.1-api>1.0.2.Final</version.org.hibernate.javax.persistence.hibernate-jpa-2.1-api>
         <version.org.hornetq>2.4.7.Final</version.org.hornetq>
         <version.org.infinispan>9.2.5.Final</version.org.infinispan>
         <version.org.jasypt>1.9.2</version.org.jasypt>
