+++ testsuite/compat/src/test/scripts/build-jars.xml
@@ -26,7 +26,7 @@
     <target name="copy-jars" description="Copy jars to be tested.">
     <copy file="${org.hibernate:hibernate-core:jar}"                       tofile="${tests.output.dir}/test-libs/hibernate-core.jar"/>
     <copy file="${org.hibernate.common:hibernate-commons-annotations:jar}" tofile="${tests.output.dir}/test-libs/hibernate-commons-annotations.jar"/>
-    <copy file="${org.hibernate:hibernate-entitymanager:jar}"              tofile="${tests.output.dir}/test-libs/hibernate-entitymanager.jar"/>
+    <copy file="${net.bytebuddy:byte-buddy:jar}"              tofile="${tests.output.dir}/test-libs/byte-buddy.jar"/>
     <copy file="${dom4j:dom4j:jar}"                                        tofile="${tests.output.dir}/test-libs/dom4j.jar"/>
     <copy file="${org.wildfly:jipijapa-hibernate5:jar}"                    tofile="${tests.output.dir}/test-libs/jipijapa-hibernate5.jar"/>
     <copy file="${antlr:antlr:jar}"                                        tofile="${tests.output.dir}/test-libs/antlr.jar"/>
