+++ testsuite/integration/basic/src/test/java/org/jboss/as/test/integration/hibernate/secondlevelcache/SFSB.java
@@ -31,13 +31,11 @@ import javax.ejb.TransactionManagementType;
 
 import org.hibernate.Session;
 import org.hibernate.SessionFactory;
-import org.hibernate.Transaction;
 import org.hibernate.cfg.AvailableSettings;
 import org.hibernate.cfg.Configuration;
 import org.hibernate.cfg.Environment;
 import org.hibernate.engine.transaction.jta.platform.internal.JBossAppServerJtaPlatform;
 import org.hibernate.internal.util.config.ConfigurationHelper;
-import org.hibernate.resource.transaction.spi.TransactionStatus;
 
 /**
  * @author Madhumita Sadhukhan
@@ -89,11 +87,15 @@ public class SFSB {
 
         try {
             Session session = sessionFactory.openSession();
-            Transaction ormTransaction = session.beginTransaction(); // join the current JTA transaction
-            TransactionStatus status = ormTransaction.getStatus();
-            if(status.isNotOneOf(TransactionStatus.ACTIVE)) {
-                throw new RuntimeException("Hibernate Transaction is not active after joining Hibernate to JTA transaction: " + status.name());
-            }
+            // Hibernate ORM 5.2+ doesn't allow beginTransaction in an active JTA transaction, as openSession
+            // will automatically join the JTA transaction.
+            // See https://github.com/hibernate/hibernate-orm/wiki/Migration-Guide---5.2
+            //Transaction ormTransaction = session.beginTransaction(); // join the current JTA transaction
+            //TransactionStatus status = ormTransaction.getStatus();
+            //if(status.isNotOneOf(TransactionStatus.ACTIVE)) {
+            //    throw new RuntimeException("Hibernate Transaction is not active after joining Hibernate to JTA transaction: " + status.name());
+            //}
+
             session.save(student);
             // trans.commit();
             session.close();
@@ -112,11 +114,14 @@ public class SFSB {
 
         try {
             Session session = sessionFactory.openSession();
-            Transaction ormTransaction = session.beginTransaction(); // join the current JTA transaction
-            TransactionStatus status = ormTransaction.getStatus();
-            if(status.isNotOneOf(TransactionStatus.ACTIVE)) {
-                throw new RuntimeException("Hibernate Transaction is not active after joining Hibernate to JTA transaction: " + status.name());
-            }
+            // Hibernate ORM 5.2+ doesn't allow beginTransaction in an active JTA transaction, as openSession
+            // will automatically join the JTA transaction.
+            // See https://github.com/hibernate/hibernate-orm/wiki/Migration-Guide---5.2
+            // Transaction ormTransaction = session.beginTransaction(); // join the current JTA transaction
+            // TransactionStatus status = ormTransaction.getStatus();
+            // if(status.isNotOneOf(TransactionStatus.ACTIVE)) {
+            //    throw new RuntimeException("Hibernate Transaction is not active after joining Hibernate to JTA transaction: " + status.name());
+            // }
             student = session.load(Student.class, id);
             session.close();
