+++ testsuite/integration/basic/src/test/java/org/jboss/as/test/integration/jpa/hibernate/sessionfactorytest/persistence.xml
@@ -29,7 +29,12 @@
             <property name="hibernate.hbm2ddl.auto" value="create-drop"/>
             <property name="hibernate.show_sql" value="false"/>
             <property name="hibernate.session_factory_name" value="modelSessionFactory"/>
-           <!-- <property name="hibernate.allow_update_outside_transaction" value="true"/> only needed for ORM 5.3.0 --> 
+            <!-- Hibernate 5.2+ (see https://hibernate.atlassian.net/browse/HHH-10877 +
+                 https://hibernate.atlassian.net/browse/HHH-12665) no longer defaults
+                 to allowing a DML operation outside of a started transaction.
+                 The application workaround is to configure new property hibernate.allow_update_outside_transaction=true.
+            -->
+           <property name="hibernate.allow_update_outside_transaction" value="true"/>
         </properties>
     </persistence-unit>
 </persistence>
