+++ testsuite/integration/basic/src/test/java/org/jboss/as/test/integration/jpa/resultstream/ResultStreamTestCase.java
@@ -20,7 +20,7 @@
  * 02110-1301 USA, or see the FSF site: http://www.fsf.org.
  */
 
-package org.wildfly.test.integration.ee8.hibernate;
+package org.jboss.as.test.integration.jpa.resultstream;
 
 import static org.junit.Assert.assertEquals;
