+++ testsuite/integration/basic/src/test/java/org/jboss/as/test/integration/jpa/resultstream/persistence.xml
