--- testsuite/integration/ee8-temp/src/test/java/org/wildfly/test/integration/ee8/hibernatesearch/persistence.xml
@@ -1,10 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<persistence xmlns="http://xmlns.jcp.org/xml/ns/persistence" version="2.1">
-    <persistence-unit name="jpa-search-test-pu">
-        <jta-data-source>java:jboss/datasources/ExampleDS</jta-data-source>
-        <properties>
-            <property name="hibernate.hbm2ddl.auto" value="create-drop"/>
-            <property name="hibernate.search.default.directory_provider" value="ram"/>
-        </properties>
-    </persistence-unit>
-</persistence>
